requirements.txt
e3nn==0.5.6
transformers==4.44.2
scikit-learn==1.5.2
tensorboard==2.17.1
numpy==1.26.4
rdkit==2025.3.5
matplotlib==3.9.2
pyyaml
timm==1.0.15
seaborn==0.13.2
tqdm==4.67.1
lmdb==1.5.1
pandas
torch-geometric

# ./appendix/pyg_lib-0.4.0+pt28cu129-cp312-cp312-linux_x86_64.whl 
# ./appendix/torch_cluster-1.6.3+pt28cu129-cp312-cp312-linux_x86_64.whl
# ./appendix/torch_scatter-2.1.2+pt28cu129-cp312-cp312-linux_x86_64.whl
# ./appendix/torch_sparse-0.6.18+pt28cu129-cp312-cp312-linux_x86_64.whl
# ./appendix/torch_spline_conv-1.2.2+pt28cu129-cp312-cp312-linux_x86_64.whl